Support #29913 - Rename sanger workflows to PCR workflows
- Changed all titles to PCR workflow.
- Nav and homepage changes to reflect new name.
- Renamed all file paths containing "sanger-workflow" to "pcr-workflow"
- Fixed tslint issues with SangerSampleSelectionStep
