+++ packages/dina-ui/components/seqdb/pcr-workflow/SangerSeqReactionStep.tsx
