+++ packages/dina-ui/components/seqdb/pcr-workflow/pcr-batch-plating-step/ContainerGrid.tsx
