+++ packages/dina-ui/components/seqdb/pcr-workflow/seq-batch-select-coordinats-step/DraggableSeqReactionBox.tsx
