+++ packages/dina-ui/pages/seqdb/sanger-workflow-sequencing/run.tsx
@@ -2,12 +2,12 @@ import { BackToListButton, LoadingSpinner } from "common-ui";
 import { PersistedResource } from "kitsu";
 import { useRouter } from "next/router";
 import PageLayout from "packages/dina-ui/components/page/PageLayout";
-import { SangerSeqReactionStep } from "packages/dina-ui/components/seqdb/sanger-workflow/SangerSeqReactionStep";
-import { SeqBatchSelectCoordinatesStep } from "packages/dina-ui/components/seqdb/sanger-workflow/seq-batch-select-coordinats-step/SeqBatchSelectCoordinatesStep";
+import { SangerSeqReactionStep } from "packages/dina-ui/components/seqdb/pcr-workflow/SangerSeqReactionStep";
+import { SeqBatchSelectCoordinatesStep } from "packages/dina-ui/components/seqdb/pcr-workflow/seq-batch-select-coordinats-step/SeqBatchSelectCoordinatesStep";
 import { useEffect, useState } from "react";
 import { Button, Spinner } from "react-bootstrap";
 import { Tab, TabList, TabPanel, Tabs } from "react-tabs";
-import { SangerSeqBatchStep } from "../../../components/seqdb/sanger-workflow/SangerSeqBatchStep";
+import { SangerSeqBatchStep } from "../../../components/seqdb/pcr-workflow/SangerSeqBatchStep";
 import { SeqdbMessage, useSeqdbIntl } from "../../../intl/seqdb-intl";
 import { SeqBatch } from "../../../types/seqdb-api";
 import { useSeqBatchQuery } from "../seq-batch/edit";
